augmentation/requirements.txt
albumentations==1.4.11
numpy==1.26.4
pillow==10.4.0
